Project setup and script to start new puzzle/day
Make this project a Deno project in VSCode.

Add a `template` directory with the prepared structure
for a new puzzle/day.

Add  `new-day.sh` to instanciate a new puzzle/day from the
template. The script can also be invoked as a Deno script:
`deno run new-day`

Git-ignore the `aoc-session-cookie.txt`.